Revert "Rename Shadowsocks.Proxy -> Shadowsocks.ForwardProxy"
This reverts commit be3cc3f3a67d63560c8e165d4e1fd55cb8e9734d.
